Move sst validation to another module
